capi/requirements.txt
bottle
pymysql
openai
python-dotenv
requests